Wait for platform to be verified on the front-end before importing site.
We should really implement a way to poll the queue remotely, but for now we just wait
a couple seconds. This should be enough time for the front-end platform verify task to
complete. Otherwise, the front-end isn't yet aware of the profiles available on the
profile, and so cannot make the reference on the site, throwing an error.