Remove 'remote_host' option from aliases.
Since setting a 'remote-host' option would cause commands to be run on the
remote, we pass in an option that won't affect operations: 'remote_host'.
We then translate that back to 'remote-host' when we're creating the alias.
However, we don't need (or want) the meaningless 'remote_host' entry, so
we remove it.