Closes #12: Strip local TLD from aliases on the host.
This has an interesting and unpleasant side-effect. Sites named without a '.'
will break in Aegir, since the context gets re-written with a 'Valkyrie' context-type.
This then fails to verify, with a 'Class Provision_Context_Valkyrie not found' error.

I believe that this is due to the verify that follows the 'provision_save' we use to
create the alias. The verify task appears to pick on the new alias's context type.
We'll have to look into this, as it could be the root of other issues.

For now, we'll just advise not to use a site name with a period in it, and probably
enforce that with an extra validation rule.